[refactor] ZTView >> ZTViewController로 리팩터링
